Refactor into radio folder
